requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.2
protobuf==3.17.3
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
